+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+1.5.0
+-----
+ * check installed nusim version
+ * install nusim
+ * nusim install Turbinado instance
+
 1.4.2
 -----
  * improve logging by day and app on day ON/OFF